@@ -44,54 +44,55 @@
     "dist": "yarn compile && electron-builder --publish never"
   },
   "dependencies": {
-    "@panter/vue-i18next": "^0.15.1",
-    "bitcoinjs-lib": "^5.1.7",
-    "bitsharesjs": "^3.1.1",
+    "@electron/remote": "^2.0.4",
+    "@panter/vue-i18next": "^0.15.2",
+    "bitcoinjs-lib": "^6.0.1",
+    "bitsharesjs": "^5.0.0",
     "bootstrap-vue": "^2.0.3",
-    "chalk": "^2.4.2",
-    "crypto-js": "^3.1.9-1",
-    "dexie": "^2.0.4",
-    "electron-devtools-installer": "^2.2.4",
+    "crypto-js": "^4.1.1",
+    "dexie": "^3.2.1",
+    "electron-devtools-installer": "^3.2.0",
     "elliptic": "^6.5.4",
-    "fs-jetpack": "^2.2.2",
-    "i18next": "^17.1.0",
-    "i18next-node-fs-backend": "^2.1.3",
+    "env": "^0.0.2",
+    "i18next": "^21.6.11",
+    "i18next-fs-backend": "^1.1.4",
     "lzma": "^2.3.2",
-    "otpauth": "^5.0.3",
+    "otpauth": "^7.0.10",
     "readline": "^1.3.0",
-    "typeface-rajdhani": "0.0.72",
-    "typeface-roboto": "0.0.75",
+    "typeface-rajdhani": "1.1.13",
+    "typeface-roboto": "1.1.13",
+    "uuid": "^8.3.2",
     "vue": "^2.6.10",
     "vue-multiselect": "^2.1.6",
     "vue-password-strength-meter": "^1.5.0",
     "vue-router": "^3.1.3",
     "vuex": "^3.1.1",
-    "ws": "^7.5.6",
+    "ws": "^8.5.0",
     "zxcvbn": "^4.4.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.6.3",
-    "@babel/plugin-proposal-object-rest-spread": "^7.6.2",
-    "@babel/plugin-transform-runtime": "^7.6.2",
-    "@babel/preset-env": "^7.6.3",
+    "@babel/core": "^7.17.5",
+    "@babel/plugin-proposal-object-rest-spread": "^7.17.3",
+    "@babel/plugin-transform-runtime": "^7.17.0",
+    "@babel/preset-env": "^7.16.11",
+    "@soda/friendly-errors-webpack-plugin": "^1.8.1",
     "babel-eslint": "^10.0.3",
-    "babel-loader": "^8.0.6",
-    "css-loader": "^3.2.0",
-    "electron": "^3.0.0",
-    "electron-builder": "^21.2.0",
-    "eslint": "^6.5.1",
-    "eslint-plugin-vue": "^5.2.3",
-    "file-loader": "^4.2.0",
-    "friendly-errors-webpack-plugin": "^1.7.0",
-    "node-sass": "^4.14.1",
-    "sass-loader": "^8.0.0",
-    "source-map-support": "^0.5.13",
-    "style-loader": "^1.0.0",
-    "vue-loader": "^15.7.1",
-    "vue-template-compiler": "^2.6.10",
-    "webpack": "^4.41.0",
-    "webpack-cli": "^3.3.9",
-    "webpack-merge": "^4.2.2",
-    "webpack-node-externals": "^1.7.2"
+    "babel-loader": "^8.2.3",
+    "css-loader": "^5.2.7",
+    "electron": "^17.0.1",
+    "electron-builder": "^22.14.13",
+    "eslint": "^8.9.0",
+    "eslint-plugin-vue": "^8.4.1",
+    "file-loader": "^6.2.0",
+    "sass": "^1.49.8",
+    "sass-loader": "^12.6.0",
+    "source-map-support": "^0.5.21",
+    "style-loader": "^3.3.1",
+    "vue-loader": "^15.9.8",
+    "vue-template-compiler": "^2.6.14",
+    "webpack": "^5.69.1",
+    "webpack-cli": "^4.9.2",
+    "webpack-merge": "^5.8.0",
+    "webpack-node-externals": "^3.0.0"
   }
 }
